app/requirements.txt
fastapi==0.101.1
uvicorn[standard]==0.22.0
motor==3.6.0
pymongo>=4.6.1,<5
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.6.0
python-dotenv==1.0.0
pydantic-settings==2.0.3
pytest==8.4.2
pytest-asyncio==0.23.7
httpx==0.25.0
